fix: emit 64 vec bug

+++ src/fstapi.c
@@ -2975,7 +2975,7 @@ void fstWriterEmitValueChangeVec32(fstWriterContext *xc,
         int br = bits & 31;
         int i;
         int w;
-        uint32_t v;
+        uint64_t v;
         unsigned char *s;
         if (FST_UNLIKELY(bits > xc->outval_alloc_siz)) {
             xc->outval_alloc_siz = bits * 2 + 1;
@@ -3020,7 +3020,7 @@ void fstWriterEmitValueChangeVec64(fstWriterContext *xc,
         int br = bits & 63;
         int i;
         int w;
-        uint32_t v;
+        uint64_t v;
         unsigned char *s;
         if (FST_UNLIKELY(bits > xc->outval_alloc_siz)) {
             xc->outval_alloc_siz = bits * 2 + 1;
